finalmente!!
finalemente o salvamento das informações do banco de dados e todas as operações matematias estão funcionando!!

+++ src/Controllers/App/Request.php
@@ -7,48 +7,13 @@ use Source\Models\UserApps;
 
 class Request
 {
-	private $router;
+	public function login($data)
+	{
+		$data = filter_var_array($data, [
+			"email" => FILTER_SANITIZE_EMAIL,
+			"password" => FILTER_SANITIZE_STRIPPED,
+		]);
 
-	public function __construct($router){
-		$this->router = $router;
-	}
-
-	public function login($data){
-
-		//echo json_encode($data);exit;	
-
-		$validateEmptyFields = array_keys($data, '');
-
-		if ($validateEmptyFields) {
-			echo json_encode(['validateEmptyFields' => $validateEmptyFields]);
-			return;
-		}
-
-		$validateFields = [];
-		$validateFields['invalidEmail'] = validateEmail($data['email']);
-
-		foreach ($validateFields as $field) {
-			if (!$field) {
-				echo json_encode(['validateFields' => $validateFields]);
-				return;
-			}
-		}
-
-		$password = base64_encode(pack('H*', sha1(utf8_encode(trim($data['password'])))));
-
-		$user = (new User)->find('password = :password and email = :email', "password={$password}&email={$data['email']}")->fetch();
-
-		if ($user) {
-			$_SESSION['user_id'] = $user->id;
-			echo json_encode('redirect');
-		} else {
-			echo json_encode(['userNotExist' => 'userNotExist']);
-		}
-	}
-
-	public function register($data){
-
-		$save = [];
 		$findEmptyFields = array_keys($data, '');
 
 		if ($findEmptyFields) {
@@ -58,35 +23,84 @@ class Request
 
 		$validateFields = [];
 
-		if(!validateEmail($data['email'])){
+		$user = (new User())->find('email = :e', "e={$data['email']}")->fetch();
+
+		if (!$user || !password_verify($data['password'], $user->password)) {
+			$validateFields['email'] = '';
+			$validateFields['password'] = 'Informações inválidas';
+		}
+
+		if ($validateFields) {
+                  echo json_encode(['validateFields' => $validateFields]);
+                  return;
+            }
+
+		if ($user) {
+			$_SESSION['user_id'] = $user->id;
+			echo json_encode('redirect');
+		}
+	}
+
+	public function register($data)
+	{
+		$save = [];
+		$findEmptyFields = array_keys($data, '');
+
+		if ($findEmptyFields) {
+			echo json_encode(['emptyFields' => $findEmptyFields]);
+			return;
+		}
+
+		$data['name'] = filter_var($data['name'], FILTER_SANITIZE_STRIPPED);
+		$data['cpf'] = filter_var($data['cpf'], FILTER_SANITIZE_STRIPPED);
+		$data['email'] = filter_var($data['email'], FILTER_SANITIZE_EMAIL);
+		$data['password'] = filter_var($data['password'], FILTER_SANITIZE_STRIPPED);
+		$data['repeat_password'] = filter_var($data['repeat_password'], FILTER_SANITIZE_STRIPPED);
+
+		/*$data = filter_var_array($data, [               // falta descobrir como aplicar esse filtro em um check box com array
+			"name" => FILTER_SANITIZE_STRIPPED,
+			"cpf" => FILTER_SANITIZE_STRIPPED,
+			"email" => FILTER_SANITIZE_EMAIL,
+			"password" => FILTER_SANITIZE_STRIPPED,
+			"repeat_password" => FILTER_SANITIZE_STRIPPED,
+			"apps" => FILTER_DEFAULT
+		]);*/
+
+		$validateFields = [];
+
+		if (!validateEmail($data['email'])) {
 			$validateFields['email'] = 'Formato de email inválido';
 		}
 
-		if((new User())->find('email = :e',"e={$data['email']}")->fetch()){
+		if ((new User())->find('email = :e', "e={$data['email']}")->fetch()) {
 			$validateFields['email'] = 'Email já foi cadastrado';
 		}
 
-		if(!validateCpf($data['cpf'])){
+		if (!validateCpf($data['cpf'])) {
 			$validateFields['cpf'] = 'Formato de CPF inválido<br>ex: 000.000.000-00';
-		}	
+		}
 
-		if((new User())->find('cpf = :c',"c={$data['cpf']}")->fetch()){
+		if ((new User())->find('cpf = :c', "c={$data['cpf']}")->fetch()) {
 			$validateFields['cpf'] = 'CPF já foi cadastrado';
 		}
 
-		if(!validateName($data['name'])){
+		if (!validateName($data['name'])) {
 			$validateFields['name'] = 'Formato do nome inválido';
+		}
+
+		if ((new User())->find('name = :name', "name={$data['name']}")->fetch()) {
+			$validateFields['name'] = 'Nome ja cadastrado';
 		}
 
 		if ($data['password'] != $data['repeat_password']) {
 			$validateFields['repeat_password'] = "Senhas não conferem";
 		}
 
-		if(empty($data['apps'])){
+		if ($data['apps'] == ['']) {
 			$validateFields['apps'] = "Escolha no mínimo um app";
 		}
 
-		if($validateFields){
+		if ($validateFields) {
 			echo json_encode(['validateFields' => $validateFields]);
 			return;
 		}
@@ -96,7 +110,7 @@ class Request
 		$user->email = $data['email'];
 		$user->cpf = $data['cpf'];
 		$user->name = $data['name'];
-		$user->password = base64_encode(pack('H*', sha1(utf8_encode(trim($data['password'])))));
+		$user->password = password_hash($data['password'], PASSWORD_DEFAULT);
 
 		$user->save();
 
@@ -107,21 +121,23 @@ class Request
 			$save[] = "user_save";
 		}
 
-		foreach ($data['apps'] as $app_id) {	
-			$userApps = new UserApps();
+		foreach ($data['apps'] as $app_id) {
+			if ($app_id != '') { // IF temporario, ate resolver a questão da filtragem do check box na linha 72
+				$userApps = new UserApps();
 
-			$userApps->user_id = $user->id;
-			$userApps->app_id = $app_id;
+				$userApps->user_id = $user->id;
+				$userApps->app_id = $app_id;
 
-			$userApps->save();
+				$userApps->save();
 
-			if ($userApps->fail()) {
-				$user->destroy();
-				echo json_encode("App {$app_id}: " . $userApps->fail()->getMessage());
-				return;
-			} else {
-				$save[] = "App{$app_id}_save";
-			}			
+				if ($userApps->fail()) {
+					$user->destroy();
+					echo json_encode("App {$app_id}: " . $userApps->fail()->getMessage());
+					return;
+				} else {
+					$save[] = "App{$app_id}_save";
+				}
+			}
 		}
 		echo json_encode('success');
 	}
